+++ packages/plugin-infra/src/dom/getAvailablePlugins.ts
