+++ packages/plugin-infra/src/dom/useActivatedPlugin.ts
