+++ packages/plugin-infra/src/entry-dom.ts
@@ -1,15 +1,2 @@
 export * from './entry.js'
-export { createInjectHooksRenderer } from './utils/createInjectHooksRenderer.js'
-export {
-    PluginI18NFieldRender,
-    type PluginI18NFieldRenderProps,
-    type PluginWrapperComponent,
-    type PluginWrapperMethods,
-    useActivatedPlugin,
-    usePluginI18NField,
-    usePluginWrapper,
-    getAvailablePlugins,
-    useActivatedPluginWeb3State,
-    useActivatedPluginWeb3UI,
-    useAllPluginsWeb3State,
-} from './hooks/index.js'
+export * from './dom/index.js'
