--- packages/plugin-infra/src/hooks/index.ts
@@ -1,7 +1,0 @@
-export * from './useActivatedPlugin.js'
-export * from './useActivatedPluginWeb3State.js'
-export * from './useActivatedPluginWeb3UI.js'
-export * from './useAllPluginsWeb3State.js'
-export * from './useI18N.js'
-
-export { type PluginWrapperMethods, type PluginWrapperComponent, usePluginWrapper } from './usePluginWrapper.js'
