+++ packages/plugin-infra/src/sns-adaptor/useSearchResultTabs.ts
@@ -2,7 +2,7 @@ import { EMPTY_LIST } from '@masknet/shared-base'
 import type { Web3Helper } from '@masknet/web3-helpers'
 import type { SearchResult } from '@masknet/web3-shared-base'
 import type { I18NStringField, Plugin } from '../entry.js'
-import { getAvailablePlugins } from '../hooks/getAvailablePlugins.js'
+import { getAvailablePlugins } from '../dom/getAvailablePlugins.js'
 
 export function useSearchResultTabs(
     definitions: Plugin.SNSAdaptor.Definition[],
